Delete old QuestActivity and fragments
