+++ app/src/main/AndroidManifest.xml
@@ -56,16 +56,6 @@
                 android:value=".MainActivity" />
         </activity>
         <activity
-            android:name=".quest.activities.QuestActivity"
-            android:label="@string/activity_quest_title"
-            android:parentActivityName=".MainActivity"
-            android:screenOrientation="portrait"
-            android:windowSoftInputMode="adjustResize">
-            <meta-data
-                android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity" />
-        </activity>
-        <activity
             android:name=".quest.activities.RepeatingQuestActivity"
             android:label="@string/title_activity_repeating_quest"
             android:parentActivityName=".MainActivity"
@@ -185,10 +175,11 @@
             android:noHistory="true"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".quest.activities.QuestDetailActivity"
+            android:name=".quest.activities.QuestActivity"
             android:label="@string/title_activity_repeating_quest"
             android:parentActivityName=".MainActivity"
-            android:screenOrientation="portrait">
+            android:screenOrientation="portrait"
+            android:windowSoftInputMode="adjustResize">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".MainActivity" />
