--- app/src/main/res/layout/fragment_quest_timer.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/root_container"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:weightSum="1">
-
-    <RelativeLayout
-        android:id="@+id/quest_details_container"
-        android:layout_width="match_parent"
-        android:layout_height="0dp"
-        android:layout_gravity="center_horizontal"
-        android:layout_weight="1">
-
-        <ProgressBar
-            android:id="@+id/quest_details_progress"
-            style="?android:attr/progressBarStyleHorizontal"
-            android:layout_width="320dp"
-            android:layout_height="320dp"
-            android:layout_centerInParent="true"
-            android:indeterminate="false"
-            android:max="100"
-            android:progress="0"
-            android:progressDrawable="@drawable/timer_progress"
-            android:secondaryProgress="100" />
-
-        <Chronometer
-            android:id="@+id/quest_details_time"
-            style="@style/TextAppearance.AppCompat.Display4"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
-            android:fontFamily="sans-serif-thin"
-            android:textColor="@color/md_light_text_87"
-            android:textSize="72sp" />
-
-    </RelativeLayout>
-
-    <RelativeLayout
-        android:layout_width="match_parent"
-        android:layout_height="72dp"
-        android:layout_marginEnd="32dp"
-        android:layout_marginStart="32dp"
-        android:layout_marginTop="16dp">
-
-        <android.support.design.widget.FloatingActionButton
-            android:id="@+id/quest_details_timer"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_centerInParent="true"
-            android:clickable="true"
-            android:scaleType="center"
-            android:src="@drawable/ic_play_arrow_white_32dp" />
-
-        <ImageButton
-            android:id="@+id/quest_details_edit"
-            android:layout_width="36dp"
-            android:layout_height="36dp"
-            android:layout_alignParentStart="true"
-            android:layout_centerVertical="true"
-            android:background="@android:color/transparent"
-            android:scaleType="fitXY"
-            android:src="@drawable/ic_create_white_24dp" />
-
-        <ImageButton
-            android:id="@+id/quest_details_done"
-            android:layout_width="36dp"
-            android:layout_height="36dp"
-            android:layout_alignParentEnd="true"
-            android:layout_centerVertical="true"
-            android:background="@android:color/transparent"
-            android:scaleType="fitXY"
-            android:src="@drawable/ic_done_white_24dp" />
-    </RelativeLayout>
-
-</LinearLayout>
