--- app/src/main/res/layout/fragment_sub_quest_list.xml
@@ -1,73 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.design.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/root_container"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@color/md_white">
-
-    <android.support.v4.widget.NestedScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <LinearLayout
-            android:id="@+id/list_container"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:focusable="true"
-            android:focusableInTouchMode="true"
-            android:orientation="vertical">
-
-            <io.ipoli.android.app.ui.EmptyStateRecyclerView
-                android:id="@+id/sub_quest_list"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:background="@color/md_white"
-                android:scrollbars="vertical"
-                android:orientation="vertical"/>
-
-            <RelativeLayout
-                android:layout_width="match_parent"
-                android:layout_height="48dp"
-                android:clickable="true"
-                android:focusable="true"
-                android:focusableInTouchMode="true">
-
-                <FrameLayout
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginStart="16dp"
-                    android:layout_centerVertical="true">
-
-                    <CheckBox
-                        android:id="@+id/dummy_checkbox_used_for_width"
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center_vertical"
-                        android:visibility="invisible" />
-
-                    <ImageView
-                        android:layout_width="wrap_content"
-                        android:layout_height="wrap_content"
-                        android:layout_gravity="center"
-                        android:src="@drawable/ic_add_accent_24dp" />
-                </FrameLayout>
-
-                <android.support.design.widget.TextInputEditText
-                    android:id="@+id/add_sub_quest"
-                    android:layout_width="match_parent"
-                    android:layout_height="match_parent"
-                    android:layout_centerVertical="true"
-                    android:layout_marginEnd="16dp"
-                    android:layout_marginStart="72dp"
-                    android:imeOptions="actionDone"
-                    android:inputType="textCapSentences"
-                    android:maxLines="1"
-                    android:text="@string/add_sub_quest"
-                    android:textColor="@color/colorAccent"
-                    android:textSize="@dimen/sub_quest_item_text_size" />
-
-            </RelativeLayout>
-
-        </LinearLayout>
-    </android.support.v4.widget.NestedScrollView>
-</android.support.design.widget.CoordinatorLayout>
